Delete Prognostic_model_training directory
Today we have removed the code for the previous version of the prognostic model training section. The uploaded code had some errors and could not be executed successfully. We have reorganized and uploaded a new version of the code for the prognostic model training section, which can replicate the results mentioned in our paper.